improvement: add an igniter installer
This allows for `mix igniter.install tidewave`. Currently the installer
must be named after the hex package itself that is being installed,
so it can't be `mix igniter.install tidewave_phoenix`, but if that is
a sticking point then I'm happy to add a way for projects to
configure a different name for their installer. Already know how it would
work if necessary.

As a result of this, I'll be able to put it up on the https://ash-hq.org
installer so users can be set up with MCP from a fresh install 🎉